# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ChangePasswordBody": {
        "example": {
          "current_password": "Exemplo@123",
          "new_password": "NovaSenha@456"
        },
        "properties": {
          "current_password": {
            "description": "Senha atual.",
            "maxLength": 256,
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "description": "Nova senha (mesma política da criação).",
            "maxLength": 256,
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "ChangePasswordBody",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Erro de domínio das rotas `/auth`, `/me` e `/admin` (ADR-008).",
        "example": {
          "error": "user_not_found",
          "message": "Usuário não encontrado."
        },
        "properties": {
          "error": {
            "description": "Slug estável do erro, ex.: `user_not_found`.",
            "title": "Error",
            "type": "string"
          },
          "message": {
            "description": "Mensagem legível, em português.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Me": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_admin": {
            "title": "Is Admin",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "profiles": {
            "items": {
              "$ref": "#/components/schemas/ProfileRef"
            },
            "title": "Profiles",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "is_admin",
          "profiles"
        ],
        "title": "Me",
        "type": "object"
      },
...
      "ProfileRef": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "is_active"
        ],
        "title": "ProfileRef",
        "type": "object"
      },
...
      "RevokeRequest": {
        "example": {
          "email": "maria@exemplo.com",
          "password": "Exemplo@123",
          "token": "<token a revogar>"
        },
        "properties": {
          "email": {
            "description": "E-mail (login) do dono do token.",
            "maxLength": 255,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "description": "Senha do usuário.",
            "maxLength": 256,
            "title": "Password",
            "type": "string"
          },
          "token": {
            "description": "Token a revogar.",
            "maxLength": 256,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "token"
        ],
        "title": "RevokeRequest",
        "type": "object"
      },
...
      "TokenRequest": {
        "example": {
          "email": "maria@exemplo.com",
          "expire_days": 90,
          "label": "claude-code",
          "password": "Exemplo@123"
        },
        "properties": {
          "email": {
            "description": "E-mail (login) do usuário.",
            "maxLength": 255,
            "title": "Email",
            "type": "string"
          },
          "expire_days": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Validade em dias; omitido usa o padrão do servidor (limitado ao máximo configurado).",
            "title": "Expire Days"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Rótulo livre para identificar o token (ex.: `claude-code`).",
            "title": "Label"
          },
          "password": {
            "description": "Senha do usuário.",
            "maxLength": 256,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "TokenRequest",
        "type": "object"
      },
      "TokenResponse": {
        "properties": {
          "expires_at": {
            "description": "Data/hora de expiração do token.",
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "token": {
            "description": "Token opaco. Exibido apenas nesta resposta; guarde-o.",
            "title": "Token",
            "type": "string"
          },
          "token_type": {
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "token",
          "expires_at"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerToken": {
        "description": "Token opaco emitido por `POST /auth/token`. Cole apenas o token (o Swagger acrescenta o prefixo `Bearer`).",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API HTTP do servidor de análise de dados via MCP: emissão de tokens (`/auth`), dados do próprio\nusuário (`/me`), administração (`/admin/*`) e `/health`.\n\n**Autenticação.** Emita um token em `POST /auth/token` (e-mail + senha), clique em **Authorize**\ne cole só o token. `/me` exige um token válido; `/admin/*` exige também administrador.\n\n**Erros.** As rotas de `/auth`, `/me` e `/admin` respondem erros de domínio no formato\n`{\"error\": \"<slug>\", \"message\": \"...\"}`. Erros de validação do corpo ou dos parâmetros vêm no\nformato padrão do FastAPI (`422`, `{\"detail\": [...]}`).\n\n**O `/mcp` não está neste documento.** O endpoint MCP (Streamable HTTP/JSON-RPC) é descrito em\n`docs/MCP.md`.\n",
    "title": "Análise de Dados Genérica com MCP — API HTTP",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/revoke": {
      "post": {
        "description": "Revoga um token do próprio usuário, provando a identidade com e-mail + senha. Não exige `Authorization`.",
        "operationId": "revoke_token_auth_revoke_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "status": "revoked"
                },
                "schema": {}
              }
            },
            "description": "Token revogado."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "E-mail ou senha inválidos (ou usuário bloqueado); mensagem sempre genérica. Slugs: `invalid_credentials`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token inexistente ou de outro usuário. Slugs: `token_not_found`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revogar token de acesso",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/token": {
      "post": {
        "description": "Troca e-mail + senha por um token opaco, usado em `Authorization: Bearer <token>` no `/mcp`, em `/me` e em `/admin/*` (botão **Authorize** do Swagger). O token só é exibido nesta resposta. `expire_days` omitido usa o padrão do servidor e não pode passar do máximo configurado. Não exige `Authorization`.",
        "operationId": "issue_token_auth_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`expire_days` acima do máximo permitido. Slugs: `invalid_expire_days`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "E-mail ou senha inválidos (ou usuário bloqueado); mensagem sempre genérica. Slugs: `invalid_credentials`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Emitir token de acesso",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/me": {
      "get": {
        "description": "Dados do usuário autenticado e seus perfis. Qualquer usuário autenticado (não exige administrador).",
        "operationId": "get_me_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `user_not_found`."
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Meus dados",
        "tags": [
          "me"
        ]
      }
    },
    "/me/password": {
      "put": {
        "description": "Troca a própria senha (exige a atual) e **revoga todos os tokens** do usuário, inclusive o usado nesta chamada.",
        "operationId": "change_password_me_password_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Requisição inválida. Slugs: `invalid_current_password`, `invalid_password`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `user_not_found`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Alterar minha senha",
        "tags": [
          "me"
        ]
      }
    }
  },
...
}